# app/Http/Controllers/AdministracionController.php
<?php

namespace App\Http\Controllers;

use App\Models\ContactoSolicitado;
use App\Models\Empresa;
use App\Models\Persona;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Validator;

class AdministracionController extends Controller
{
    /**
     * @OA\Get(
     *     path="/admin/contactos",
     *     operationId="getContactosSolicitados",
     *     tags={"Administración"},
     *     summary="Listar contactos solicitados",
     *     @OA\Parameter(name="estado", in="query", required=false,
     *         @OA\Schema(type="string",
     *             enum={"pendiente","contactado","entrevista","seleccionado","no-seleccionado","proceso-cerrado"})),
     *     @OA\Response(response=200, description="Listado exitoso",
     *         @OA\JsonContent(type="array", @OA\Items(ref="#/components/schemas/ContactoSolicitado")))
     * )
     */
    public function listarContactos(Request $request): JsonResponse
    {
        $query = ContactoSolicitado::with(['empresa', 'persona']);
        if ($request->has('estado')) {
            $query->where('estado', $request->input('estado'));
        }
        return $this->successResponse($query->orderBy('created_at', 'desc')->get());
    }

    /**
     * @OA\Post(
     *     path="/admin/contactos",
     *     operationId="crearContactoSolicitado",
     *     tags={"Administración"},
     *     summary="Registrar solicitud de contacto",
     *     description="Una empresa solicita contactar a un talento. No puede existir una solicitud activa previa.",
     *     @OA\RequestBody(required=true,
     *         @OA\JsonContent(ref="#/components/schemas/ContactoSolicitadoInput")),
     *     @OA\Response(response=201, description="Contacto registrado",
     *         @OA\JsonContent(ref="#/components/schemas/ContactoSolicitado")),
     *     @OA\Response(response=409, description="Ya existe una solicitud activa"),
     *     @OA\Response(response=422, description="Errores de validación")
     * )
     */
    public function crearContacto(Request $request): JsonResponse
    {
        $validator = Validator::make($request->all(), [
            'empresa_id'  => 'required|exists:empresas,id',
            'persona_id'  => 'required|exists:personas,id',
            'notas_admin' => 'nullable|string',
        ]);

        if ($validator->fails()) {
            return $this->errorResponse('Los datos enviados no son válidos.', 422, $validator->errors()->toArray());
        }

        $existente = ContactoSolicitado::where('empresa_id', $request->empresa_id)
            ->where('persona_id', $request->persona_id)
            ->whereNotIn('estado', ['no-seleccionado', 'proceso-cerrado'])
            ->first();

        if ($existente) {
            return $this->errorResponse('Ya existe una solicitud activa entre esta empresa y talento.', 409);
        }

        $contacto = ContactoSolicitado::create($validator->validated());
        return $this->successResponse($contacto->load(['empresa', 'persona']), 201);
    }

    /**
     * @OA\Patch(
     *     path="/admin/contactos/{id}/estado",
     *     operationId="actualizarEstadoContacto",
     *     tags={"Administración"},
     *     summary="Actualizar estado de contacto",
     *     description="Cambia el estado del proceso. Las fechas se registran automáticamente según el estado.",
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="string", format="uuid")),
     *     @OA\RequestBody(required=true,
     *         @OA\JsonContent(
     *             required={"estado"},
     *             @OA\Property(property="estado", type="string",
     *                 enum={"pendiente","contactado","entrevista","seleccionado","no-seleccionado","proceso-cerrado"}),
     *             @OA\Property(property="notas_admin", type="string", nullable=true)
     *         )
     *     ),
     *     @OA\Response(response=200, description="Estado actualizado",
     *         @OA\JsonContent(ref="#/components/schemas/ContactoSolicitado")),
     *     @OA\Response(response=404, description="Contacto no encontrado")
     * )
     */
    public function actualizarEstado(Request $request, string $contacto): JsonResponse
    {
        $model = ContactoSolicitado::find($contacto);
        if (!$model) {
            return $this->errorResponse('Contacto no encontrado.', 404);
        }

        $validator = Validator::make($request->all(), [
            'estado'      => 'required|in:pendiente,contactado,entrevista,seleccionado,no-seleccionado,proceso-cerrado',
            'notas_admin' => 'nullable|string',
        ]);

        if ($validator->fails()) {
            return $this->errorResponse('Los datos enviados no son válidos.', 422, $validator->errors()->toArray());
        }

        $data = $validator->validated();

        if ($data['estado'] === 'contactado' && !$model->fecha_contacto) {
            $data['fecha_contacto'] = now();
        } elseif ($data['estado'] === 'entrevista' && !$model->fecha_entrevista) {
            $data['fecha_entrevista'] = now();
        } elseif (in_array($data['estado'], ['seleccionado', 'no-seleccionado']) && !$model->fecha_resultado) {
            $data['fecha_resultado'] = now();
        }

        $model->update($data);
        return $this->successResponse($model->load(['empresa', 'persona']));
    }

    /**
     * @OA\Get(
     *     path="/admin/estadisticas",
     *     operationId="getEstadisticas",
     *     tags={"Administración"},
     *     summary="Estadísticas generales de la plataforma",
     *     @OA\Response(response=200, description="Estadísticas generadas",
     *         @OA\JsonContent(
     *             @OA\Property(property="total_personas",       type="integer", example=45),
     *             @OA\Property(property="personas_validadas",   type="integer", example=38),
     *             @OA\Property(property="total_empresas",       type="integer", example=12),
     *             @OA\Property(property="empresas_validadas",   type="integer", example=10),
     *             @OA\Property(property="contactos_pendientes", type="integer", example=5),
     *             @OA\Property(property="contactos_en_proceso", type="integer", example=8),
     *             @OA\Property(property="contactos_exitosos",   type="integer", example=15)
     *         )
     *     )
     * )
     */
    public function estadisticas(): JsonResponse
    {
        $data = cache()->remember('estadisticas', 600, function () {
            return [
                'total_personas'       => Persona::count(),
                'personas_validadas'   => Persona::where('validado', true)->count(),
                'total_empresas'       => Empresa::count(),
                'empresas_validadas'   => Empresa::where('validado', true)->count(),
                'contactos_pendientes' => ContactoSolicitado::where('estado', 'pendiente')->count(),
                'contactos_en_proceso' => ContactoSolicitado::whereIn('estado', ['contactado', 'entrevista'])->count(),
                'contactos_exitosos'   => ContactoSolicitado::where('estado', 'seleccionado')->count(),
            ];
    });

        return $this->successResponse($data);
    }
}