# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.\n\n## Rate Limiting\nLa API implementa limitación de solicitudes para garantizar disponibilidad:\n- **Endpoints públicos:** 60 solicitudes por minuto\n- **Endpoints administrativos:** 30 solicitudes por minuto\n\n## Caché\nAlgunos endpoints utilizan caché para mejorar el rendimiento:\n- **GET /personas:** caché de 5 minutos\n- **GET /empresas:** caché de 5 minutos\n- **GET /admin/estadisticas:** caché de 10 minutos\n\n## Tiempos de respuesta esperados\n- Operaciones de lectura: < 200ms\n- Operaciones de escritura: < 500ms",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/admin/contactos": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Listar contactos solicitados",
                "operationId": "getContactosSolicitados",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "contactado",
                                "entrevista",
                                "seleccionado",
                                "no-seleccionado",
                                "proceso-cerrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ContactoSolicitado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Administración"
                ],
                "summary": "Registrar solicitud de contacto",
                "description": "Una empresa solicita contactar a un talento. No puede existir una solicitud activa previa.",
                "operationId": "crearContactoSolicitado",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactoSolicitadoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactoSolicitado"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ya existe una solicitud activa"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/admin/contactos/{id}/estado": {
            "patch": {
                "tags": [
                    "Administración"
                ],
                "summary": "Actualizar estado de contacto",
                "description": "Cambia el estado del proceso. Las fechas se registran automáticamente según el estado.",
                "operationId": "actualizarEstadoContacto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "contactado",
                                            "entrevista",
                                            "seleccionado",
                                            "no-seleccionado",
                                            "proceso-cerrado"
                                        ]
                                    },
                                    "notas_admin": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContactoSolicitado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactoSolicitadoInput": {
                "title": "Contacto Solicitado Input",
                "description": "Datos para registrar una solicitud de contacto",
                "required": [
                    "empresa_id",
                    "persona_id"
                ],
                "properties": {
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "notas_admin": {
                        "type": "string",
                        "example": "Empresa interesada en perfil para cargo de analista",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "empresa_id": "550e8400-e29b-41d4-a716-446655440001",
                    "persona_id": "550e8400-e29b-41d4-a716-446655440000",
                    "notas_admin": "Empresa interesada en perfil para cargo de analista"
                }
            },
            "ContactoSolicitado": {
                "title": "Contacto Solicitado",
                "description": "Solicitud de contacto entre empresa y talento",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "contactado",
                            "entrevista",
                            "seleccionado",
                            "no-seleccionado",
                            "proceso-cerrado"
                        ],
                        "example": "entrevista"
                    },
                    "notas_admin": {
                        "type": "string",
                        "example": "Candidato destacado para el cargo",
                        "nullable": true
                    },
                    "fecha_contacto": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-01T10:00:00+00:00",
                        "nullable": true
                    },
                    "fecha_entrevista": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-10T15:00:00+00:00",
                        "nullable": true
                    },
                    "fecha_resultado": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-30T09:00:00+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-10T15:00:00+00:00"
                    }
                },
                "type": "object",
                "example": {
                    "id": "550e8400-e29b-41d4-a716-446655440002",
                    "empresa_id": "550e8400-e29b-41d4-a716-446655440001",
                    "persona_id": "550e8400-e29b-41d4-a716-446655440000",
                    "estado": "entrevista",
                    "notas_admin": "Candidato destacado para el cargo de desarrollador",
                    "fecha_contacto": "2026-05-01T10:00:00+00:00",
                    "fecha_entrevista": "2026-05-10T15:00:00+00:00",
                    "fecha_resultado": null,
                    "created_at": "2026-04-30T09:00:00+00:00",
                    "updated_at": "2026-05-10T15:00:00+00:00"
                }
            },
...
        }
    },
...
}